updated controllers and added tests

+++ app/controller/notes.controller.js
@@ -2,9 +2,13 @@ const NotesService = require('../services/notes.service');
 
 async function getNotes(req, res) {
     try {
-        const notes = await NotesService.getAll(req.body.lead_id);
+        if (req.body.lead_id) {
+            const notes = await NotesService.getAll(req.body.lead_id);
 
-        return res.status(200).json({ status: 'success', message: 'All notes', notes });
+            return res.status(200).json({ status: 'success', message: 'All notes', notes });
+        }
+
+        return res.status(400).json({ status: 'error' });
     } catch (err) {
         res.status(400).json({ status: 'error', message: 'Server Error!' });
         throw err;
@@ -19,9 +23,13 @@ async function createNote(req, res) {
     }
 
     try {
-        const note = await NotesService.create(note_param);
+        if (note_param.user_id && note_param.lead_id && note_param.message && note_param.message != '') {
+            const note = await NotesService.create(note_param);
 
-        return res.status(201).json({ status: 'success', message: 'Note created!', note });
+            return res.status(201).json({ status: 'success', message: 'Note created!', note });
+        }
+
+        return res.status(400).json({ status: 'error' });
     } catch (err) {
         res.status(400).json({ status: 'error', message: 'Server Error!' });
         throw err;
@@ -30,9 +38,13 @@ async function createNote(req, res) {
 
 async function deleteNote(req, res) {
     try {
-        const deleted = await NotesService.delete(req.body.note_id);
+        if (req.body.note_id) {
+            const deleted = await NotesService.delete(req.body.note_id);
 
-        return res.status(200).json({ status: 'success', message: 'Note deleted!', deleted });
+            return res.status(200).json({ status: 'success', message: 'Note deleted!', deleted });
+        }
+
+        return res.status(400).json({ status: 'error' });
     } catch (err) {
         res.status(400).json({ status: 'error', message: 'Server Error!' });
         throw err;
